Adds remove method to ConfigFile. Integrates into MessageConfig
Adds __set__ command to MessageConfigs

Summary:
This revision enhances MessageConfigs.

From now on, instead of writing directly to the values in the table,
the protocol is to make a request on a reserved __set__ topic
(much like __save__ and __revert__)

This enhanced protocol improves logic and syncronization on the robot side which allows
for clients to remove values, which was the ultimate task.
Fixes T39

Test Plan:
Numerous tests done on testbed
__set__ method properly sets, adds, and removes as protocol requires

Maniphest Tasks: T39

Differential Revision: http://phabricator.team3637.com/D9